+++ src/index.ts
@@ -1,3 +1,5 @@
+'use client';
+
 export { default as Live2DWidget } from './Live2DWidget';
 export type { Live2DWidgetProps, ModelName } from './Live2DWidget';
 export { BUILT_IN_MODELS, getRandomModel } from './Live2DWidget';