+++ test/types/react-syntax-highlighter.d.ts
@@ -1,2 +1,13 @@
 declare module 'react-syntax-highlighter';
 declare module 'react-syntax-highlighter/dist/esm/styles/prism';
+declare module '*.css' {
+  const css: string;
+  export default css;
+}
+
+declare module '*.scss' {
+  const css: string;
+  export default css;
+}
+
+declare module './globals.css';